+++ home.php
@@ -20,9 +20,14 @@
     if (($has_left === TRUE) and ($has_right === TRUE)):
       echo '<div class="col-sm-3 col-md-2 col-lg-2" role="complementary">';
 
+      if (is_front_page() && wrdsb_athome()) {}
+
       if (is_front_page() && wrdsb_i_am_a_school_exception()) { ?>
+
+        <p><a href="https://schools.wrdsb.ca/athome/" onclick="ga('send','event','schoolBanners', 'click_banner','WRDSB@Home', 'https://schools.wrdsb.ca/athome/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="http://schools.wrdsb.ca/athome/files/2020/03/Website-Button.jpg" alt="WRDSB@Home" /></a></p>
+        
         <p><a href="https://www.wrdsb.ca/blog/tags/coronavirus/" onclick="ga('send','event','schoolBanners', 'click_banner','COVID-19', 'https://www.wrdsb.ca/blog/tags/coronavirus/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="http://schools.wrdsb.ca/wp-content/uploads/2020/03/COVID-Website-Button.jpg" alt="COVID-19 Information" /></a></p>
-        <p><a href="https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/" onclick="ga('send','event','schoolBanners', 'click_banner','severe weather', 'https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://wrdsb-ui-assets.s3.amazonaws.com/public/2/2.2.0/images/Severe+Weather+Information%40344.png" alt="Severe Weather Information" /></a></p>
+        <!--p><a href="https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/" onclick="ga('send','event','schoolBanners', 'click_banner','severe weather', 'https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://wrdsb-ui-assets.s3.amazonaws.com/public/2/2.2.0/images/Severe+Weather+Information%40344.png" alt="Severe Weather Information" /></a></p-->
         <p><a href="https://www.wrdsb.ca/labour/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'labour', 'https://www.wrdsb.ca/labour/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://www.wrdsb.ca/wp-content/uploads/Labour-Update.png" alt="Labour Updates" /></a></p>
         <p><a href="https://myway.wrdsb.ca/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'myway', 'https://myway.wrdsb.ca/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://www.wrdsb.ca/wp-content/uploads/myway_banner_344x100.jpg" alt="MyWay Logo"/></a></p>
       <?php 
@@ -47,8 +52,9 @@
       }
 
       if (is_front_page() && wrdsb_i_am_a_school()) { ?>
+        <p><a href="https://schools.wrdsb.ca/athome/" onclick="ga('send','event','schoolBanners', 'click_banner','WRDSB@Home', 'https://schools.wrdsb.ca/athome/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="http://schools.wrdsb.ca/athome/files/2020/03/Website-Button.jpg" alt="WRDSB@Home" /></a></p>
         <p><a href="https://www.wrdsb.ca/blog/tags/coronavirus/" onclick="ga('send','event','schoolBanners', 'click_banner','COVID-19', 'https://www.wrdsb.ca/blog/tags/coronavirus/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="http://schools.wrdsb.ca/wp-content/uploads/2020/03/COVID-Website-Button.jpg" alt="COVID-19 Information" /></a></p>
-        <p><a href="https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/" onclick="ga('send','event','schoolBanners', 'click_banner','severe weather', 'https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://wrdsb-ui-assets.s3.amazonaws.com/public/2/2.2.0/images/Severe+Weather+Information%40344.png" alt="Severe Weather Information" /></a></p>
+        <!--p><a href="https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/" onclick="ga('send','event','schoolBanners', 'click_banner','severe weather', 'https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://wrdsb-ui-assets.s3.amazonaws.com/public/2/2.2.0/images/Severe+Weather+Information%40344.png" alt="Severe Weather Information" /></a></p-->
         <p><a href="https://www.wrdsb.ca/labour/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'labour', 'https://www.wrdsb.ca/labour/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://www.wrdsb.ca/wp-content/uploads/Labour-Update.png" alt="Labour Updates" /></a></p>
         <p><a href="https://myway.wrdsb.ca/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'myway'" target="_blank" rel="noopener"><img src="https://www.wrdsb.ca/wp-content/uploads/myway_banner_344x100.jpg" alt="MyWay Logo"/></a></p>
         <p><a href="https://schools.wrdsb.ca/school-year-information/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'syi', '/about/school-year-information',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://www.wrdsb.ca/wp-content/uploads/schoolyearinformation_344x100.jpg" alt="School Year Information" target="_blank" rel="noopener" /></a></p>
@@ -60,8 +66,10 @@
 
       get_sidebar('left');
 
-
+    if (is_front_page() && wrdsb_athome()) {}
+    
     if (is_front_page() && wrdsb_i_am_a_school_exception()) { ?>
+        <p><a href="https://schools.wrdsb.ca/athome/" onclick="ga('send','event','schoolBanners', 'click_banner','WRDSB@Home', 'https://schools.wrdsb.ca/athome/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="http://schools.wrdsb.ca/athome/files/2020/03/Website-Button.jpg" alt="WRDSB@Home" /></a></p>
          <p><a href="https://www.wrdsb.ca/blog/tags/coronavirus/" onclick="ga('send','event','schoolBanners', 'click_banner','COVID-19', 'https://www.wrdsb.ca/blog/tags/coronavirus/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="http://schools.wrdsb.ca/wp-content/uploads/2020/03/COVID-Website-Button.jpg" alt="COVID-19 Information" /></a></p>
        <p><a href="http://www.wefihelps.org/" target="_blank" rel="noopener" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'wefi', 'http://www.wefihelps.org/',{'nonInteraction':1});"><img src="https://wrdsb-ui-assets.s3.amazonaws.com/public/2/2.0.0/images/wefi.png" alt="Waterloo Education Foundation Inc. (WEFI)"></a></p>
      <?php 
@@ -74,9 +82,12 @@
     elseif (($has_left === TRUE) and ($has_right === FALSE)):
       echo '<div class="col-sm-3 col-lg-3" role="complementary">';
 
+      if (is_front_page() && wrdsb_athome()) {}
+
       if (is_front_page() && wrdsb_i_am_a_school_exception()) { ?>
+        <p><a href="https://schools.wrdsb.ca/athome/" onclick="ga('send','event','schoolBanners', 'click_banner','WRDSB@Home', 'https://schools.wrdsb.ca/athome/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="http://schools.wrdsb.ca/athome/files/2020/03/Website-Button.jpg" alt="WRDSB@Home" /></a></p>
         <p><a href="https://www.wrdsb.ca/blog/tags/coronavirus/" onclick="ga('send','event','schoolBanners', 'click_banner','COVID-19', 'https://www.wrdsb.ca/blog/tags/coronavirus/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="http://schools.wrdsb.ca/wp-content/uploads/2020/03/COVID-Website-Button.jpg" alt="COVID-19 Information" /></a></p>
-        <p><a href="https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/" onclick="ga('send','event','schoolBanners', 'click_banner','severe weather', 'https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://wrdsb-ui-assets.s3.amazonaws.com/public/2/2.2.0/images/Severe+Weather+Information%40344.png" alt="Severe Weather Information" /></a></p>
+        <!--p><a href="https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/" onclick="ga('send','event','schoolBanners', 'click_banner','severe weather', 'https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://wrdsb-ui-assets.s3.amazonaws.com/public/2/2.2.0/images/Severe+Weather+Information%40344.png" alt="Severe Weather Information" /></a></p-->
         <p><a href="https://www.wrdsb.ca/labour/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'labour', 'https://www.wrdsb.ca/labour/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://www.wrdsb.ca/wp-content/uploads/Labour-Update.png" alt="Labour Updates" /></a></p>
         <p><a href="https://myway.wrdsb.ca/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'myway', 'https://myway.wrdsb.ca/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://www.wrdsb.ca/wp-content/uploads/myway_banner_344x100.jpg" alt="MyWay Logo"/></a></p>
       <?php 
@@ -101,8 +112,9 @@
       }
 
       if (is_front_page() && wrdsb_i_am_a_school()) { ?>
+        <p><a href="https://schools.wrdsb.ca/athome/" onclick="ga('send','event','schoolBanners', 'click_banner','WRDSB@Home', 'https://schools.wrdsb.ca/athome/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="http://schools.wrdsb.ca/athome/files/2020/03/Website-Button.jpg" alt="WRDSB@Home" /></a></p>
          <p><a href="https://www.wrdsb.ca/blog/tags/coronavirus/" onclick="ga('send','event','schoolBanners', 'click_banner','COVID-19', 'https://www.wrdsb.ca/blog/tags/coronavirus/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="http://schools.wrdsb.ca/wp-content/uploads/2020/03/COVID-Website-Button.jpg" alt="COVID-19 Information" /></a></p>
-       <p><a href="https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/" onclick="ga('send','event','schoolBanners', 'click_banner','severe weather', 'https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://wrdsb-ui-assets.s3.amazonaws.com/public/2/2.2.0/images/Severe+Weather+Information%40344.png" alt="Severe Weather Information" /></a></p>
+       <!--p><a href="https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/" onclick="ga('send','event','schoolBanners', 'click_banner','severe weather', 'https://www.wrdsb.ca/our-schools/student-transportation/severe-weather/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://wrdsb-ui-assets.s3.amazonaws.com/public/2/2.2.0/images/Severe+Weather+Information%40344.png" alt="Severe Weather Information" /></a></p-->
         <p><a href="https://www.wrdsb.ca/labour/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'labour', 'https://www.wrdsb.ca/labour/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://www.wrdsb.ca/wp-content/uploads/Labour-Update.png" alt="Labour Updates" /></a></p>
         <p><a href="https://myway.wrdsb.ca/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'myway'" target="_blank" rel="noopener"><img src="https://www.wrdsb.ca/wp-content/uploads/myway_banner_344x100.jpg" alt="MyWay Logo"/></a></p>
         <p><a href="https://schools.wrdsb.ca/school-year-information/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'syi', '/about/school-year-information',{'nonInteraction':1});"><img src="https://www.wrdsb.ca/wp-content/uploads/schoolyearinformation_344x100.jpg" alt="School Year Information" target="_blank" rel="noopener" /></a></p>
@@ -114,8 +126,9 @@
 
       get_sidebar('left');
     
+      if (is_front_page() && wrdsb_athome()) {}
+
       if (is_front_page() && wrdsb_i_am_a_school_exception()) { ?>
-        <p><a href="https://www.wrdsb.ca/blog/tags/coronavirus/" onclick="ga('send','event','schoolBanners', 'click_banner','COVID-19', 'https://www.wrdsb.ca/blog/tags/coronavirus/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="http://schools.wrdsb.ca/wp-content/uploads/2020/03/COVID-Website-Button.jpg" alt="COVID-19 Information" /></a></p>
         <p><a href="http://www.wefihelps.org/" onclick="ga('send', 'event', 'schoolBanners', 'click_banner', 'wefi', 'http://www.wefihelps.org/',{'nonInteraction':1});" target="_blank" rel="noopener"><img src="https://wrdsb-ui-assets.s3.amazonaws.com/public/2/2.0.0/images/wefi.png" alt="Waterloo Education Foundation Inc. (WEFI)"></a></p>
       <?php 
       }