+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.5.0
++ added wrdsb_athome() function
++ added WRDSB@Home button to school sites
+- removed severe weather button from school sites for now
+
 ## 2.4.2
 + added GTM plugin function